requirements.txt
fastapi 
uvicorn 
python-multipart
python-dotenv
aiofiles
pydantic
pydantic-settings
httpx
python-jose
passlib